Move reconciler.go into operations
