+++ contrib/mesos/pkg/scheduler/errors/errors.go
@@ -18,9 +18,11 @@ package errors
 
 import (
 	"errors"
+	"fmt"
 )
 
 var (
 	NoSuchPodErr  = errors.New("No such pod exists")
 	NoSuchTaskErr = errors.New("No such task exists")
+	ReconciliationCancelledErr = fmt.Errorf("explicit task reconciliation cancelled")
 )
